Rework exports to be async
