+++ app/controllers/series_controller.rb
@@ -1,4 +1,6 @@
 class SeriesController < ApplicationController
+  include ExportHelper
+
   before_action :set_series, except: %i[index new create indianio_download]
 
   before_action :check_token, only: %i[show overview]
@@ -138,7 +140,7 @@ class SeriesController < ApplicationController
       user = User.find_by(email: email)
       if user
         options = { deadline: true, only_last_submission: true, with_info: true, all_students: true, indianio: true }
-        send_zip Export.new(item: @series, list: @series.exercises, users: [user], options: options).bundle
+        send_zip Zipper.new(item: @series, list: @series.exercises, users: [user], options: options).bundle
       else
         render json: { errors: ['Unknown email'] }, status: :not_found
       end
@@ -222,5 +224,13 @@ class SeriesController < ApplicationController
     raise Pundit::NotAuthorizedError if @series.hidden? &&
                                         !current_user&.course_admin?(@series.course) &&
                                         @series.access_token != params[:token]
+  end
+
+  def send_zip(zip)
+    send_data zip[:data],
+              type: 'application/zip',
+              filename: zip[:filename],
+              disposition: 'attachment',
+              x_sendfile: true
   end
 end
