+++ app/models/user.rb
@@ -39,6 +39,7 @@ class User < ApplicationRecord
   has_many :repository_admins, dependent: :restrict_with_error
   has_many :courses, through: :course_memberships
   has_many :events, dependent: :restrict_with_error
+  has_many :exports, dependent: :destroy
 
   has_many :subscribed_courses,
            lambda {
