+++ app/policies/course_policy.rb
@@ -131,6 +131,12 @@ class CoursePolicy < ApplicationPolicy
     show?
   end
 
+  def export?
+    return true if zeus?
+
+    course_member?
+  end
+
   def permitted_attributes
     # record is the Course class on create
     if course_admin? || (record == Course && user&.admin?)
@@ -143,6 +149,10 @@ class CoursePolicy < ApplicationPolicy
   private
 
   def course_admin?
+    record.class == Course && user&.course_admin?(record)
+  end
+
+  def course_member?
     record.class == Course && user&.course_admin?(record)
   end
 end
