+++ app/policies/series_policy.rb
@@ -77,16 +77,18 @@ class SeriesPolicy < ApplicationPolicy
     course_admin?
   end
 
-  def scoresheet_download?
-    course_admin?
-  end
-
   def mass_rejudge?
     course_admin?
   end
 
   def reset_token?
     edit?
+  end
+
+  def export?
+    return true if zeus?
+
+    course_member?
   end
 
   def permitted_attributes
@@ -102,5 +104,9 @@ class SeriesPolicy < ApplicationPolicy
 
   def course_admin?
     record.class == Series && user&.course_admin?(record&.course)
+  end
+
+  def course_member?
+    record.class == Series && user&.member_of?(record&.course)
   end
 end
