+++ app/policies/user_policy.rb
@@ -77,6 +77,12 @@ class UserPolicy < ApplicationPolicy
     show?
   end
 
+  def export?
+    return true if zeus?
+
+    record == user
+  end
+
   def permitted_attributes
     if user&.zeus?
       %i[username first_name last_name email permission time_zone]
