+++ app/views/exports/download_submissions.html.erb
@@ -62,11 +62,11 @@
                 </div>
                 <div class="field form-group">
                   <div class="col-sm-6 col-sm-offset-3">
-                      <input type="radio" id="only_last_submission" name="last_submission" checked value="1">
+                      <input type="radio" id="only_last_submission" name="only_last_submission" checked value="1">
                       <label class="checkbox-label" for="only_last_submission"><%= t '.only_last_submission' %>
                   </div>
                   <div class="col-sm-6 col-sm-offset-3">
-                      <input type="radio" id="not_only_last_submissions" name="last_submission" value="0">
+                      <input type="radio" id="not_only_last_submissions" name="only_last_submission" value="0">
                       <label class="checkbox-label" for="not_only_last_submissions"><%= t '.all_submissions' %>
                   </div>
                 </div>
