+++ app/views/layouts/_navbar.html.erb
@@ -36,6 +36,9 @@
               <li class="dropdown-header"><%= current_user.full_name %></li>
               <li><%= link_to t("layout.menu.profile"), current_user %></li>
               <li><%= link_to t("layout.menu.my_submissions"), submissions_path(user_id: current_user) %></li>
+              <% if Export.where(user: current_user).exists? %>
+                <li><%= link_to t("layout.menu.my_exports"), exports_path %></li>
+              <% end %>
               <% if policy(:pages).toggle_demo_mode? || policy(:pages).toggle_dark_mode? %>
                 <li role="separator" class="divider"></li>
               <% end %>
