+++ app/views/users/show.html.erb
@@ -101,7 +101,7 @@
             </a>
             <ul class="dropdown-menu dropdown-menu-right">
               <li>
-                <%= link_to t('.download_my_submissions'), export_user_path(@user) %>
+                <%= link_to t('.download_my_submissions'), for_user_exports_path(@user) %>
               </li>
             </ul>
           <% end %>
