+++ config/deploy/production.rb
@@ -12,26 +12,25 @@ server 'tityos.ugent.be',   user: 'dodona', port: '4840', roles: %w[app worker]
 
 set :branch, 'master'
 
-set :delayed_job_pools_per_server, {
+set :delayed_job_pools_per_server,
     'dodona' => {
-        'default,statistics' => 2,
+      'default,statistics,exports,cleaning' => 2
     },
     'sisyphus' => {
-        'submissions,low_priority_submissions,high_priority_submissions' => 6,
+      'submissions,low_priority_submissions,high_priority_submissions' => 6
     },
     'salmoneus' => {
-        'submissions,low_priority_submissions,high_priority_submissions' => 6,
+      'submissions,low_priority_submissions,high_priority_submissions' => 6
     },
     'tantalus' => {
-        'submissions,low_priority_submissions,high_priority_submissions' => 6,
+      'submissions,low_priority_submissions,high_priority_submissions' => 6
     },
     'ixion' => {
-        'submissions,low_priority_submissions,high_priority_submissions' => 6,
+      'submissions,low_priority_submissions,high_priority_submissions' => 6
     },
     'tityos' => {
-        'submissions,low_priority_submissions,high_priority_submissions' => 6,
+      'submissions,low_priority_submissions,high_priority_submissions' => 6
     }
-}
 
 # server 'example.com', user: 'deploy', roles: %w{app web}, other_property: :other_value
 # server 'db.example.com', user: 'deploy', roles: %w{db}
