+++ config/locales/views/defaults/en.yml
@@ -29,6 +29,7 @@ en:
       manual: User manual
       exercises: Exercises
       my_submissions: My submissions
+      my_exports: My exports
       submissions: Submissions
       users: Users
       courses: Courses
