+++ config/routes.rb
@@ -1,9 +1,8 @@
 Rails.application.routes.draw do
-  devise_for :users, controllers: {omniauth_callbacks: 'omniauth_callbacks'}
+  devise_for :users, controllers: { omniauth_callbacks: 'omniauth_callbacks' }
   root 'pages#home'
 
-
-  authenticated :user, -> user {user.zeus?} do
+  authenticated :user, ->(user) { user.zeus? } do
     mount DelayedJobWeb, at: '/dj'
   end
 
@@ -25,7 +24,7 @@ Rails.application.routes.draw do
     concern :mediable do
       member do
         constraints host: Rails.configuration.default_host do
-          get 'media/*media', to: 'exercises#media', constraints: {media: /.*/}, as: 'media'
+          get 'media/*media', to: 'exercises#media', constraints: { media: /.*/ }, as: 'media'
         end
       end
     end
@@ -40,7 +39,7 @@ Rails.application.routes.draw do
       resources :submissions, only: %i[index create]
     end
 
-    resources :series, except: [:new, :index] do
+    resources :series, except: %i[new index] do
       resources :exercises, only: [:index]
       member do
         get 'available_exercises', to: 'exercises#available'
@@ -55,28 +54,28 @@ Rails.application.routes.draw do
     end
     get 'series/indianio/:token', to: 'series#indianio_download', as: 'indianio_download'
 
-    scope '/export' do
-      get '/series/:id', to: 'export#download_submissions_from_series', as: 'export_series'
-      post '/series/:id', to: 'export#start_download_from_series'
-      get '/courses/:id', to: 'export#download_submissions_from_course', as: 'export_course'
-      post '/courses/:id', to: 'export#start_download_from_course'
-      get '/users/:id', to: 'export#download_submissions_from_user', as: 'export_user'
-      post '/users/:id', to: 'export#start_download_from_user'
+    resources :exports, except: %i[show edit update new destroy create] do
+      get 'for_user/:id', on: :collection, to: 'exports#new_for_user', as: 'for_user'
+      post 'for_user/:id', on: :collection, to: 'exports#create_for_user'
+      get 'for_course/:id', on: :collection, to: 'exports#new_for_course', as: 'for_course'
+      post 'for_course/:id', on: :collection, to: 'exports#create_for_course'
+      get 'for_series/:id', on: :collection, to: 'exports#new_for_series', as: 'for_series'
+      post 'for_series/:id', on: :collection, to: 'exports#create_for_series'
     end
 
     resources :courses do
-      resources :series, only: [:new, :index] do
-        resources :exercises, only: [:show, :edit, :update], concerns: %i[mediable submitable infoable]
+      resources :series, only: %i[new index] do
+        resources :exercises, only: %i[show edit update], concerns: %i[mediable submitable infoable]
       end
-      resources :exercises, only: [:show, :edit, :update], concerns: %i[mediable submitable infoable]
+      resources :exercises, only: %i[show edit update], concerns: %i[mediable submitable infoable]
       resources :submissions, only: [:index]
-      resources :members, only: [:index, :show, :edit, :update], controller: :course_members do
+      resources :members, only: %i[index show edit update], controller: :course_members do
         get 'download_labels_csv', on: :collection
         post 'upload_labels_csv', on: :collection
       end
       member do
         get 'statistics'
-        get 'subscribe/:secret', to: 'courses#registration', as: "registration"
+        get 'subscribe/:secret', to: 'courses#registration', as: 'registration'
         get 'manage_series'
         get 'scoresheet'
         post 'mass_accept_pending'
@@ -98,7 +97,7 @@ Rails.application.routes.draw do
             root to: 'exercises#description', as: 'description'
             get 'media/*media',
                 to: 'exercises#media',
-                constraints: {media: /.*/},
+                constraints: { media: /.*/ },
                 as: 'description_media'
           end
         end
@@ -130,7 +129,7 @@ Rails.application.routes.draw do
       member do
         get 'download'
         get 'evaluate'
-        get 'media/*media', to: 'submissions#media', constraints: {media: /.*/}, as: 'media'
+        get 'media/*media', to: 'submissions#media', constraints: { media: /.*/ }, as: 'media'
       end
     end
 
@@ -148,7 +147,7 @@ Rails.application.routes.draw do
     resources :labels
     resources :programming_languages
 
-    resources :institutions, only: [:index, :show, :edit, :update]
+    resources :institutions, only: %i[index show edit update]
     resources :events, only: [:index]
 
     scope 'stats', controller: 'statistics' do
