+++ test/factories/exports.rb
@@ -1,0 +1,5 @@
+FactoryBot.define do
+  factory :export do
+    user { nil }
+  end
+end
